populate fundamentals error

+++ go/sql/schema/20260531033826_create_equities_table.sql
@@ -3,9 +3,9 @@ CREATE TABLE equity_universe (
     conid INT PRIMARY KEY,
     ticker VARCHAR(12) NOT NULL,
     exchange VARCHAR(12) NOT NULL,
-    pe_ratio NUMERIC(12, 4),
-    price_to_sales NUMERIC(12, 4),
-    return_on_equity NUMERIC(12, 4),
+    pe_ratio DOUBLE PRECISION,
+    price_to_sales DOUBLE PRECISION,
+    return_on_equity DOUBLE PRECISION,
     business_summary TEXT,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
